The classes of the system was completed
For this step were modified the class Bowling_Game.hpp and Bowling_Game.cpp.
The methods startGame, eval_Inputs and addScore were added.
The first one is in charge to starting the simulation of the bowling game.
The second is responsible to evaluate the different cases that may occur in
this game. And finally the addScore method, it generates the score obtained through a strike or a spare.